Update workflow to avoid duplication of URL in rabbit MQ
We have changed the duplication check process from storing a successfully 
crawled URL to storing enqueued URLs. This allows us to block URLs from 
being inserted into the queue multiple times. This will reduce the size 
of the rabbit queue, reduce instances of duplicate processing and hopefully 
avoid disk issues - due to the queue being persisted to disk. This will 
may result in a larger memory footprint for redis as it will need to store 
more values for longer.

In addition to this we have removed the wait period after a document is 
crawled, which blocked it being reinserted into the queue. This is because 
the delay between enqueuing and processing is long enough to negate the 
benefit of this addition delay.